+++ packages/component/src/block/positions/sections/limitConfirm.tsx
@@ -16,16 +16,18 @@ interface Props {
 export const LimitConfirm: FC<Props> = (props) => {
   const { order, quote, side } = props;
 
+  console.log("LimitConfirm::::::::", props);
+
   return (
     <div>
       <div className="orderly-space-y-3">
         <div className="orderly-font-medium orderly-text-base-contrast-54 orderly-text-2xs desktop:orderly-text-sm">
-          {`You will close ${order.order_quantity} ETH position at limit price.`}
+          {`You will close ${order.order_quantity} ${props.base} position at limit price.`}
         </div>
       </div>
       <Divider className="orderly-my-4" />
       <div className="orderly-mb-4 orderly-flex orderly-text-base orderly-items-center orderly-gap-2">
-        <NetworkImage size={20} type={"token"} name={"ETH"} />
+        <NetworkImage size={20} type={"token"} name={props.base} />
         <Text
           rule="symbol"
           className="orderly-text-base desktop:orderly-text-lg"