+++ packages/walletconnect/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @orderly.network/web3-modal
 
+## 0.0.2-alpha.7
+
+### Patch Changes
+
+- fix bug
+
+## 0.0.2-alpha.6
+
+### Patch Changes
+
+- fix bug
+
+## 0.0.2-alpha.5
+
+### Patch Changes
+
+- fix bug
+- fix bug
+- fix bug
+
+## 0.0.2-alpha.4
+
+### Patch Changes
+
+- fix bugs
+- fix bugs
+
+## 0.0.2-alpha.3
+
+### Patch Changes
+
+- fix bug
+
 ## 0.0.2-alpha.6
 
 ### Patch Changes